+++ packages/subgraph/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ### Unreleased
+
+- Add entities/events/handler for DepositManager `DepositInstantiated` event